Refactoring/MoveComponentsToPageContent
