+++ components/cards/cafeCard/cafeCard.module.scss
@@ -11,7 +11,7 @@
   height: 130px;
   border-radius: 20px;
 
-  justify-content: end;
+  justify-content: flex-end;
   background-repeat:no-repeat;
   background-position: center center;
   background-size: 100%;
