+++ pages-content/auth/auth.module.scss
@@ -99,7 +99,7 @@
   background: white;
 
   &:active {
-    background: $bg-young-night;
+    background: $grey-rainy;
   }
 }
