+++ shared/url-manager.ts
@@ -1,15 +1,7 @@
-import { log } from 'util';
-
 class URLManagerStore {
-  public getHomeURL() {
-    return '/';
-  }
-  public getLoginURL() {
-    return '/auth/login';
-  }
-  public getRegistrationURL() {
-    return '/auth/registration';
-  }
+  public getHomeURL() {return '/';}
+  public getLoginURL() {return '/auth/login';}
+  public getRegistrationURL() {return '/auth/registration';}
 }
 
 const URLManager = new URLManagerStore();
