--- shared/urls.ts
@@ -1,11 +1,0 @@
-export const UrlManager: any = {
-  home: '/',
-  login: '/auth/login',
-  registration: '/auth/registration',
-};
-
-// class URLManagerStore {
-//   public getHomeURL = () => '/';
-//   public getLoginURL = () => '/';
-//   public getRegistrationURL = () => '/';
-// }
